rtem-hackathon/pyproject.toml
[project]
name = "rtem-hackathon"
version = "0.1.0"
description = "Tools to help with the RTEM Hackathon."
authors = ["danlsn <danlsn@github.com>"]

[tool.poetry]
name = "rtem-hackathon"
version = "0.1.0"
description = "Tools to help with the RTEM Hackathon."
authors = ["danlsn <danlsn@github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"onboard.client" = "^1.8.0"
pandas = "^1.4.2"
tqdm = "^4.64.0"
requests-cache = "^0.9.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
